Add support to redirect large tests to civ2 shared-runners for test_models.py (#1576)
### Ticket
#1738 

### Problem description
- Some models hit OOM on CiV1 runners for wormhole, need to be directed
to CiV2 shared runners

### What's changed
- Update infra for ability to tag test_models tests needing more host
memory with "large" for CiV2 runners
- Pass --arch using original arch before mapping to civ2 runs-on name
via generate_test_matrix.py
- Turns out all 6 of the host OOM tests won't fit on n150 anyways and
are 7B-14B param models, so just updated failing reason to be consistent
with other models that are too big for single chip. Add new job in the
future when this change is really needed.

### Checklist
- [x] Tested on branch, tagged "large" and new job redirects to CiV2:
https://github.com/tenstorrent/tt-xla/actions/runs/18379954774/job/52363924444