Build docker image on push to main (#1732)
### Ticket

[Slack](https://tenstorrent.slack.com/archives/C08DJEGTNCR/p1760818006493579)

### Problem description
"Push to main" workflows expect that docker image already exists, but
this is not always true.
If we have multiple changes to docker structure pushed through different
PRs, main can have a new docker tag that is not yet built.

### What's changed
Update logic in "Push to main" workflows to allow for docker build if it
is missing.

Jobs in "test_forge_models_push" were running regardless on errors, also
updated that to not run if docker image is missing @kmabeeTT

### Checklist
- [ ] New/Existing tests provide coverage for changes