modify CI + uplift mlir (#1669)
### Ticket
Closes https://github.com/tenstorrent/tt-xla/issues/1668

### Problem description
A test hangs, which is blocking an uplift. There are important changes
in tt-mlir we want uplifted ASAP.

Edit: In the meantime a lot was figured out, TLDR is that pytest
--forked kills processes and that completely messes up cleanup in metal,
which leads to hangs.

### What's changed
~~Skipped a test.~~ Edit: In the meantime a lot was figured out, TLDR is
that we will A) not fork multichip tests B) run singlechip tests
exclusively on n150 to avert a hang, not an ideal solution but it is
what it is
Uplifted mlir

### Checklist
- [ ] New/Existing tests provide coverage for changes