Run nightly torch multidevice tests without forked (#1748)
This is caught in the blast radius of #1669, where the UB we were
creating by leaking metal singleton resources due to forked process
abnormal exit started causing hangs.

Currently, forked tests run on multidevice configurations will hang on
the second test while acquiring a device and the system will be in a
hung state.

Also, any subsequent tests in future processes will also hang until the
device is reset, but this is not caught in CI since we only launch a
single main pytest process when unforked.

- [x] Test run of forge-models-passing for the two changed testgroups is
here: https://github.com/tenstorrent/tt-xla/actions/runs/18693692143 ->
These tests do not fully pass due to unrelated regressions. However,
they do not hang either, which is the point of this PR.