[tt-torch] experimental compile flow
Current compile flow produces pytorch tracing (and possibly `torch_xla`)
overhead on each execution of the forward call (of the compiled model).

High-level overview (current compile flow):

On each forward call, the model graph is fully traced and after that
the `sync` operation is performed on the output tensors. `torch_xla`
then figures out the graph and looks up for its executable in its cache.
If it can find one, it will execute it, otherwise it will compile the
graph, store its executable in the cache and execute it. On subsequent
executions of the `forward` call it will reuse the previously created
cache entry and avoid compiling the graph again.

To avoid these overheads, I experimented with implementing a similar
flow as the official `openxla` backend in torch. This turned out to be
rather complex, since there were a lot of test failures that were caused
by this - hangs, sharding issues, etc. Because of that, we've decided to
go with a subset of the changes under a compile option, to allow further
work on debugging and implementing a more complete backend compile flow,
as well as enabling perf team to continue investigating further
bottlenecks.

The experimental compile flow is triggered with:
```torch.compile(model, inputs,
options={"tt_experimental_compile": True})```

For now, experimental compile leverages `extract_compiled_graph()` from
the `torch_xla.core.dynamo_bridge`. This function basically replaces the
graph module with a module containing a single "op" called
`optimized_mod`. This op will handle executing the cached graph(s)
directly - without tracing overhead.

To enable using `optimized_mod` we need to inline (lift) all of the
module's arguments (inputs, params and consts) into the function
signature. This is done via torch export API.

The experimental flow overview:
- if we have an already compiled graph - just run it
- otherwise, export the module (lift the arguments), populate all the
  parameters/constants and save them for the subsequent execution,
  create the compiled graph via `extract_compiled_graph()` and execute
  it.