Add semi manual multichip PoC tests in pytorch (#1735)
Closes #1024
Semi-manual means that inputs and weights sharding is annotated by the
user manually (torch_xla.runtime.mark_sharding), but compiler is
propagating sharding annotations from inputs and weights through the
model and inserting collective ops.

What is changed:
- Added data_parallel_inference_generic that does data parallel
inference when given model, input and batch dimension. This can be used
for every model that needs to be run in Data Parallel mode in Pytorch.
- Added ccl op tests in pytorch for AllReduce and AllGather
- Added MNIST manual data parallel test.
- Added MNIST manual tensor parallel with xfail on mlir
[issue](https://github.com/tenstorrent/tt-mlir/issues/5290).
All multichip tests are added only for n300 setup.