Update Training Test Error Messages (#1965)
### Ticket
N/A

### Problem description
Few MLIR fixes have landed, so this PR is updating error messages for
training tests for visibility. Also, it is hard to combine inference job
CIs in a query due to different namings of the jobs.

### What's changed
This PR includes:
- Rename of a tag `@pytest.mark.training` to
`@pytest.mark.test_forge_models_training`, to align training CI job
names with inference
- Update training preset to respect newly added tag
- Fix error messages for models