[pjrt] ensure client destruction on process exit (#1999)
`torch_xla` doesn't call `PJRT_ClientDestroy` properly. This means that
we are not closing the devices properly.

Recently, this started causing hangs on `n300` boards on subsequent
execution of tests.

This PR introduces a global singleton object which will ensure that we
are properly destroying the client instance on process shutdown. The
singleton serves as a fallback mechanism if the framework doesn't call
`PJRT_ClientDestroy` - like in the case of `torch_xla`.

Additionally, optimizer sub-meshes are now closed after each
compilation; this previously was needed to avoid hangs, but now it
causes them. Leaving the mechanism of persisting optimizer submesh in
the code base, so that we can play with it if needed. Obviously, we need
to dig deep into these issues to fix them properly.

NOTE: this does not solve the case when the process terminates abruptly,
e.g. in case of `SIGSEGV` (segmentation fault). For this, ideally we
would want a fix on `tt-metal` side.

Closes #1824