[infra] stop cleaning cache on non-CI environments
The code which checks if the various cache dirs (huggingface, jax, etc.)
is wrong - it detects regular user environments outside of IRD as CIv2.

Fixing this by adding `TT_XLA_CI` env variable when running tests on CI.
So that we can make sure that the cleanup only happens when we are
actually running on CI.

Additionally, cleaning up the code to avoid double negations.