Model `tt.paged_update_cache` as custom op (#2016)
### Ticket
N/A
### Problem description
`ttnn.paged_update_cache` is now modelled in tt-mlir, will be included
in next uplift.

This op is required for implementing paged attention/cache in the vllm
plugin.

### What's changed
Model `tt.paged_update_cache` as a custom torch op with underlying
stablehlo custom call. Model CPU implementation of the op for testing.

Add a unit test suite for the op

### Checklist
- [X] New/Existing tests provide coverage for changes