Add dynamic tracking of BringupStatus for test infra (#2012)
### Ticket
#1343

### Problem description
explained in #1343 and #1052

### What's changed
This PR is solving the part where we want to dynamicly check for bringup
status of models, without reading from `config.yaml`-s. I also change a
bit how some functions work to be able to add this change.

How it works:
We create file `_bringup_stage.txt` at the start of test using
`pytest.fixture` and deleting it at the end of the single test using the
same fixture. We write to that file using `LOG_BRINGUP_STAGE` API
implemented for logging the current bringup stage (we override the file
every time we hit different bringup stage). Later on, we just read the
last checkpoint we reached and classify it as one of the `BringupStatus`
enums.

Left the logic for us to be able to write `reasons` and `bringup_status`
for models that segfault the CI, so we need to mark them as
`ModelTestStatus.NOT_SUPPORTED_SKIP`.

I have also remove `capteesys` from `test_all_models` becuase there is
no need for it after i deleted `update_test_metadata_for_exception`
function (the function we used prevouslly for detecting runtime
bringup_status)

To enable this feature, use `ENABLE_BRINGUP_STAGE_LOGGING=1` env var
inside nightly CI.