[vLLM] Add BGE-M3 tests and vllm v0.10.1.1 patch (#2095)
### Ticket
N/A

### Problem description
We want to test bge-m3 and its performance through vLLM. This requires a
few minor changes to the vllm_tt files and a patch containing a
[bugfix](https://github.com/vllm-project/vllm/commit/faa7a5daac8244376a3a182d6eee8d2e0f6d8127)
for vllm v0.10.1.1.

### What's changed
Added BGE-M3 as a example in examples/vllm and added a test and perf
test in tests/integrations on push.


We now install vllm from
`https://github.com/tenstorrent/vllm.git@forge/v0.10.1.1_patched` which
is equivalent to v0.10.1.1 with this
[bugfix](https://github.com/vllm-project/vllm/commit/faa7a5daac8244376a3a182d6eee8d2e0f6d8127)
(needed to run encoder_only models). This is temporary as the bugfix is
included in v0.11.0 which can be used once we support torch version >=
v2.8.0.


integrations/vllm/vllm_tt now allows encode and encode_only models to
run. Block table calculations are skipped if there's no kv_cache to
avoid invalid index errors.

[EDIT]: `VLLM_TARGET_DEVICE="empty"` set in venv/activate to avoid error when building vllm.

### Checklist
- [x] New/Existing tests provide coverage for changes