Change file caching env variables behaviour in CI (#2236)
### Ticket
/

### Problem description
File fetching logic in tt-forge-models requires the env variables
`DOCKER_CACHE_ROOT` and `IRD_LF_CACHE` to not be defined at all in order
to work properly.

### What's changed
`DOCKER_CACHE_ROOT` is not defined at all instead of being defined with
an empty value in the specific case.
`IRD_LF_CACHE` is only defined when running the job on a Civ2 runner.

### Checklist
- [ ] New/Existing tests provide coverage for changes