Add load_config for models with graph testing (#281)
### Ticket
https://github.com/tenstorrent/tt-xla/issues/2153

### Problem description
Currently our graph tests in tt-xla load the entire model and its
pre-trained weights when only a small part of the model is needed
(attention, mlp, etc.). An improvement that allows these tests to run
faster is to expose the config in model loader and then load the desired
module (attention, mlp, etc.) with that config.

### What's changed
Added `load_config()` as an optional function in base.py ForgeModel.
Defined `load_config()` for models currently tested in the torch graph
tests in tt-xla.

### Checklist
- [x] New/Existing tests provide coverage for changes