Add mapping for the new n300-high-memory label (#2246)
### Ticket
Slack

### Problem description
We have 2 flavours of n300 runners, one with 32GB of memory and one with
48GB of memory.
Developers need to be able to specify which one they want to use for
their tests because more memory is sometimes needed.

### What's changed
Introduce the new allowed label value and map it to a CIv2 label
variant.

Note: I've also added the n300-high-memory label on adequate machines.

### Checklist
- [ ] New/Existing tests provide coverage for changes