CopyFromBuffer Device2Device BufferInstance transfer workaround  (#2277)
### Ticket
Revealed by #2258.

### Problem description
After #1657, output BufferInstances no longer have a host runtime
tensor. If we try to do a (PJRT) device to device transfer (even though
that is not really meaningful in the current way we model
MeshDevice/PJRT Device Instance), the copyToBuffer path assumes an
existing host runtime tensor on the source buffer instance.

### What's changed
In copyToBuffer, if a device runtime tensor exists on the copy source
buffer instance, transfer it to host as the source of truth.

### Checklist
- [x] New/Existing tests provide coverage for changes