Add `--forked` to Weekly Workflow and Move Large Tests to CIv2 (#2331)
### Ticket
N/A

### Problem description
Due to the problem in naming, where we had an extra `torch` in the name
of the jobs, weekly jobs were not run with `--forked` option. Apart from
that, weekly training tests include large tests, so we need to make sure
we run them on shared runners so that we have enough RAM.

### What's changed
- Add `run_forge_models_torch` to the list of jobs that need `--forked`.
- Switch `shared-runners` option to true for large training weekly
workload.