[PJRT] Initial changes for PJRT unit/mock tests (#2320)
### Ticket

N/A

### Problem description

PJRT API needs a framework for executing tests in isolation, in addition
to existing e2e integration workflows (Python tests). The idea is to use
this framework to create unit and mock tests which will put PJRT code in
the focus of testing. This is the initial PR which introduces the
necessary changes for building and running tests, and some example unit
tests for PJRT utils. The problem of mocking PJRT dependencies (tt-mlir)
will be solved in a future PR (this will also make it easier to review
this one).

### What's changed

Read full description of new changes in the new `tests/pjrt/README.md`.

### Best way to review

0. New README.md
1. Top-level CMakeLists.txt, then new leaf CMakeLists.txt
2. Tests

### What is **not** included in this PR

- Meaningful unit / mock tests, apart from example unit tests that
validate PJRT helpers, and can be used as an example of how to add new
tests.
- Changes that integrate new tests with CI. This will be done in a
different PR, after more tests have been added.

### Checklist

- [x] New/Existing tests provide coverage for changes