Nightly fail inspector (#2403)
### Ticket
https://github.com/tenstorrent/tt-xla/issues/1419

### Problem description
Bisect commits for failed nightly tests looking for the first bad
result.

### What's changed
Workflow is set to run automatically after failed nightly. In this mode
it will send results to the slack.

However, it can be also run with manual dispatch. In such case one
should provide **nightly's** run id. There is also an option to provide
test filter by test name (eg. if single or specific test(s) need to be
bisected) but this test has to fail in selected nightly and not to fail
in previous nightly. There is also an option to run bisect multiple
number of times useful for bisect of flaky tests.

### Checklist
- [ ] New/Existing tests provide coverage for changes