Add pytorch TP example (#2326)
### Ticket
https://github.com/tenstorrent/tt-xla/issues/2318

### Problem description
There are currently no simple examples for how to run a pytorch model in
tensor-parallel.

### What's changed
Added qwen3_tp.py in examples/pytorch that shows how to load a model
from transformers and run it in tensor parallel. Key points for running
in tensor parallel are:
- Set SPMD mode
- Create a mesh
- Shard chosen model layers/modules along `"model"` dimension

[EDIT]: added an overview of tensor parallelism (TP) strategy for LLM
MLP and attention layers with graphs to help visualize.

### Checklist
- [x] New/Existing tests provide coverage for changes