Fix deadlock in concurrent copyToHost, add testcases (#2324)
### Ticket
Fixes a deadlock spotted by @pilkicTT in #2125 

### Problem description
There is an obvious deadlock in the implementation of #2125 after
changing to unique_lock, and the exact mechanism is discussed in the
comments of #2125.

### What's changed
- Use separate locks for:
- (Instance lock) Preventing multiple framework threads from
concurrently joining on the same copyToHost thread for the **same**
BufferInstance, which is C++ UB
- (Static lock) Preventing multiple framework threads from concurrently
running runtime toHost() between **different** BufferInstances, due to
metal's lack of op-dispatch thread safety. This is locked from within
the async copyToHost thread
- Reset BufferInstance m_copy_to_host_thread after joining. This is a
separate bug, where a later framework thread could join on an
already-destructed thread
- Add testcases for (1) concurrent copyToHost calls on the same
bufferInstance and (2) concurrent copyToHost calls on multiple
bufferInstances

### Checklist
- [x] New/Existing tests provide coverage for changes