Add PCC enable/threshold guidance to junit xml tags for model tests used downstream (#2486)
### Ticket
https://github.com/tenstorrent/tt-xla/issues/2484

### Problem description
- It is currently difficult to visually spot tests that can have pcc
check enabled or pcc threshold raised
- Was locally having this logic to detect in scripts processing pcc from
artifacts, but it is better to compute it at the source

### What's changed
- Add logic to compute guidance tags for two specific cases (enabling
pcc check, and raising pcc threshold) that can be prioritized for
automatic updates. It can be read from summarize_junit_xmls.py to start,
next which is being added via
https://github.com/tenstorrent/tt-xla/pull/2475, and soon will be the
basis of automatic test promotion script/flow.

### Checklist
- [x] Tested locally in handful of cases, will include results in PR
comments.