Fix Weights Tying Problem in TT-XLA Torch Infra (#2527)
### Ticket
NA

### Problem description
Currently many training tests are failing on frontend issues related to
missing gradients. After investigation it was noticed that
`transformers` library uses feature of tied weights. This feature is
well supported for CPU evaluation, but when it comes to TT, since we are
using torch_xla, it is faulty. The problem is that torch_xla unties
weights on model movement to device.

### What's changed
This PR introduces following changes:
- Call to `tie_weights` after moving to device for both CPU and TT. This
is to ensure both runs properly use these weights.
- Update of training tests that were failing because of the problem with
tied weights (`ttir.gather` was out of date, it was failing with the
same problem as other tests).

### Checklist
- [x] Torch test infra fix