Changes needed for torch 2.9.0 and transformers 4.57.1 uplifts (#338)
### Ticket
#1020 in tt-xla

### Problem description
tt-xla is updating the transformers library to 4.57.1 and torch to
2.9.0, we need to make changes on how we load some of the transformers
models based on the updated in the library since our last update

### What's changed
1. Transformers changed how they process text/image inputs for Fuyu
models, so we don't longer need to `model_inputs["image_patches"][0]`
2. Transformers changed the location of the `ALL_PARALLEL_STYLES` global
variable
3. New transformers cache logic causing SEGFAULT issues if the model
uses sliding window attention. We need to match sliding window value to
input size so it does not go out of bounds when updating the cache.
4. Force padding to `right` by default.
5. _supports_sdpa is not set by default before `__init__` so we need to
add property for OpenVLA loader
6. whisper requirements uplifted to torchcodec==0.8 to match
compatibility with torch==2.9.0

### Checklist
- [x] New/Existing tests provide coverage for changes