Adding support for device compile option in jax.jit (#1300)
Currently, due to our runtime executing code on devices starting with
index zero, jax threw an error if the user tried to manually set a
device with non-zero index through `jax.jit`, due to a mismatch.
Changing the PJRT compile so that we extract the device from the
`compile_options` protobuf, and assign the relevant devices to the
`ExecutableInstance`. Please note that we still **run code from devices
starting with index 0**, this change merely aligns what the jax
libraries sees so that we do not have a mismatch.

Also see issue https://github.com/tenstorrent/tt-xla/issues/938