Install g++-12 in base Docker image (#2563)
### Ticket

https://github.com/tenstorrent/tt-xla/issues/2487

### Problem description

The base Docker image doesn't come with version 12 implementation of
`libstdc++`, only with version 11. IRD and CI derived images have both
11 and 12, and during the build process `clang` chooses version 12.
Builds have started to fail in containers using the base image because
of specific C++ features used in `tt-metal` (see linked issue), which do
not compile on version 11. Since `tt-metal` declares the usage of C++ 20
standard, and version 12 is a more comprehensive implementation of C++
20, I am including it in the base image to ensure the project compiles
everywhere.

I haven't been able yet to figure out why IRD and CI containers have
version 12. It doesn't come from the image build process directly, nor
indirectly as a dependency of some other package (I recreated the IRD
image from base image step-by-step). I also suspected that there may be
a volume mapping between host and container that enabled the IRD
container to access the version 12 headers on the host, but that is not
the case. I plan to follow up on this.

### What's changed

The `g++-12` package, and as a dependency `gcc-12`, are now installed
during the base Docker image build process.