Skip Long Running Training Test (#2644)
### Ticket
N/A

### Problem description
We have a test that is running for 3 hours for a single forward +
backward pass. This is from the family of segformers where we already
skipped tests, so we need to fix this. Link to the job:
https://github.com/tenstorrent/tt-xla/actions/runs/20391513033/job/58601367371

### What's changed
Skipping the test for now until we figure out what the problem with the
model is.