[Torch/XLA] Decompose torch.dot as torch.matmul (#2681)
## Issue

https://github.com/tenstorrent/tt-xla/issues/2672

## PR description

A default decomposition of `torch.dot` results in `stablehlo.multiply`
without reduction in the final StableHLO graph when `tt_torch` backend
is used, resulting in incorrect results of `torch.dot` calls (result is
equivalent to eltwise multiply). A custom decomposition is added that
transforms `torch.dot` to `torch.matmul` for 1D tensors. This is just a
temporary workaround - we will push for a fix to be made in the default
decomposition table for the `dot` op.