Op by op mode test (#2494)
### Ticket
https://github.com/tenstorrent/tt-xla/issues/1002
Solution from this PR has drawbacks explained in [issue
comment](https://github.com/tenstorrent/tt-xla/issues/1002#issuecomment-3627583883);
it is temporary until a solution that is appropriate for both CI and
local testing is implemented.

### Problem description
We want to have a test that tests models in op-by-op mode, which should:
- run the model to get the StableHLO IR;
- wrap each op in its own module;
- compile and run each op module;
- create a report with results.

### What's changed
- added `--dump-irs` option to `test_all_models_torch/jax` methods ->
used instead of `--serialize` because it dumps IR modules even if
compilation didn't finish successfully (we use StableHLO IR for op_by_op
for now and existence of it is enough)
- removed `full` from configs/parametrization/test_id creation ->
op_by_op is just another mode of testing, it doesn't need to have its
own config
For example, if we collect tests with 
`pytest --collect-only -q tests/runner/test_models.py -m
"expected_passing and n150"`
we will get op_by_op tests for models that are expected to pass in
regular (full) mode.
- extracted shared logic for JAX and Torch tests to reduce code
duplication
- added test_all_models_op_by_op method that can be called to test a
model in op by op mode

For this test to work:
1. project needs to be built with `-DTTMLIR_ENABLE_BINDINGS_PYTHON=ON`
2. system descriptor needs to be saved using `ttrt query
--save-artifacts`
3. for the report to be created, test should be called with
`--json-report --json-report-file=report.json`

### Notes
For now, only single chip models were considered.
If this test is ran on N300 machine, it might leave it in inconsistent
state, needing a reset.

### Checklist
- [ ] New/Existing tests provide coverage for changes